+++ json_serializable/lib/src/shared_checkers.dart
@@ -3,11 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/element/type.dart';
-import 'package:collection/collection.dart';
 import 'package:source_gen/source_gen.dart' show TypeChecker;
+import 'package:source_helper/source_helper.dart';
 
 import 'helper_core.dart';
-import 'utils.dart';
 
 /// A [TypeChecker] for [Iterable].
 const coreIterableTypeChecker = TypeChecker.fromUrl('dart:core#Iterable');
@@ -20,17 +19,7 @@ const coreMapTypeChecker = TypeChecker.fromUrl('dart:core#Map');
 ///
 /// If [type] does not extend [Iterable], an error is thrown.
 DartType coreIterableGenericType(DartType type) =>
-    typeArgumentsOf(type, coreIterableTypeChecker).single;
-
-/// If [type] is the [Type] or implements the [Type] represented by [checker],
-/// returns the generic arguments to the [checker] [Type] if there are any.
-///
-/// If the [checker] [Type] doesn't have generic arguments, `null` is returned.
-List<DartType> typeArgumentsOf(DartType type, TypeChecker checker) {
-  final implementation = _getImplementationType(type, checker) as InterfaceType;
-
-  return implementation.typeArguments;
-}
+    type.typeArgumentsOf(coreIterableTypeChecker)!.single;
 
 /// A [TypeChecker] for [String], [bool] and [num].
 const simpleJsonTypeChecker = TypeChecker.any([
@@ -54,7 +43,7 @@ String asStatement(DartType type) {
   }
 
   if (coreMapTypeChecker.isAssignableFromType(type)) {
-    final args = typeArgumentsOf(type, coreMapTypeChecker);
+    final args = type.typeArgumentsOf(coreMapTypeChecker)!;
     assert(args.length == 2);
 
     if (args.every((e) => e.isLikeDynamic)) {
@@ -65,6 +54,3 @@ String asStatement(DartType type) {
   final typeCode = typeToCode(type);
   return ' as $typeCode';
 }
-
-DartType? _getImplementationType(DartType type, TypeChecker checker) =>
-    type.typeImplementations.firstWhereOrNull(checker.isExactlyType);