+++ json_serializable/lib/type_helper.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/shared_checkers.dart' show simpleJsonTypeChecker, typeArgumentsOf;
+export 'src/shared_checkers.dart' show simpleJsonTypeChecker;
 export 'src/type_helper.dart'
     show TypeHelperContext, TypeHelperContextWithConfig, TypeHelper;
 export 'src/type_helpers/big_int_helper.dart';